pyproject.toml
[project]
name = "moderation-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.7",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "cache>=1.0.3",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.10",
    "punq>=0.7.0",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "uvicorn>=0.36.0",
]
